download: move download into its own package so its available for reuse
